auth_service/requirements.txt
Flask
ldap3
PyJWT
python-dotenv
PyYAML
prometheus_client
